move errors to a property to improve performance
