+++ src/users/components/organisation-access-permissions/organisation-access-permissions.component.html
@@ -17,7 +17,7 @@
     >
     </lib-gov-checkbox>
 
-    <ng-container *ngIf="enableCaseManagement">
+    <ng-container>
       <app-solicitor-profile-content
         *ngIf="hasSolicitorProfile"
       ></app-solicitor-profile-content>
