+++ src/users/models/basic-access-types.model.ts
@@ -1,0 +1,6 @@
+export interface BasicAccessTypes {
+  isPuiUserManager: boolean;
+  isPuiOrganisationManager: boolean;
+  isPuiFinanceManager: boolean;
+  isCaseAccessAdmin: boolean;
+}
