+++ src/users/models/case-management-permissions.model.ts
@@ -1,0 +1,6 @@
+import { UserAccessType } from '@hmcts/rpx-xui-common-lib';
+
+export interface CaseManagementPermissions {
+  manageCases: boolean;
+  userAccessTypes: UserAccessType[];
+}
