+++ src/users/models/personal-details.model.ts
@@ -1,0 +1,5 @@
+export interface PersonalDetails {
+  firstName: string;
+  lastName: string;
+  email: string;
+}
